Batch job output parsing improvements.
Parse the output of the batch job incrementally rather
than waiting until the entire batch is finished.
For any jobs that we know are complete (i.e. they
echoed the success token) put them in the finished
queue so the output can be reported.

Ensure we only run one batch subprocess at once.