Fix the case where batch process exits early.
Now that we run DBS without -k by default, any error
cause that process to quit before echoing all the job IDs.

In that case we weren't considering them finished, which
resulted in the builder calling DoWork() with no Subprocess running,
which would hang.

Now determine all the jobs that we haven't received status tokens
for as the set of (all jobs) - (jobs we have output for).
Put those into the completed_jobs_ vector for processing.